feat(dev): fallback to siteEntity data (#35)
Use the entity data if it is available. If not, fallback to the
siteEntity.

We will later want to add support for layouts. Layout data will be
chosen if the entityData is not present.

entity -> layout -> siteEntity

Verified that data is saved to entity for 'individual' role and to the
site for 'global' role.