feat: Migrate from QueryParams to postMessage (#51)
This coincides with the Gerrit CR in the comments. Migrated sending data
to the editor to postMessage. This changes the editor logic to pull data
from the correct source. It also sends data to the VE as needed and
changes how we handle clearing local changes in the Puck to tell VE we
did so.

Note: This is breaking without the Gerrit CR also going through