send correct data to parent
this commit changes how we watch for changes and what data we send up. We used to send just the puck data and the wrong hash. Now we send the current puck history incl. data, ui, id and the history-id as the hash. This will help in loading data to puck later