feat: add phone component (#158)
This creates a new Phone component that takes a string value from
"type.phone" entity fields. If the number is a US phone number (or the
same format as a US phone number), it will be reformatted into the
following structure, with the country code being included if it exists
on the entity field: +1 (123) 456-7890.