get path logic and rename template (#193)
Logic to return a preset path based on whether or not the entity has an
address or not and ignore the slug field. We will likely eventually go
back to the slug field, but for now, we want to avoid path collisions or
users have to think about an additional field. I thought I had included
this in one of the component polish/starter improvements user story but
I guess I missed it.

I also renamed the location template to "main" since it can technically
be used for any entity/page type.

---------

Co-authored-by: Matt Kilpatrick <mkilpatrick@yext.com>